refactor and remove duplicated classes
