+++ CoreImportExportPlugins/src/au/gov/asd/tac/constellation/plugins/importexport/ImportTableColumn.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package au.gov.asd.tac.constellation.plugins.importexport.delimited;
+package au.gov.asd.tac.constellation.plugins.importexport;
 
 import au.gov.asd.tac.constellation.graph.attribute.AttributeDescription;
 import au.gov.asd.tac.constellation.graph.attribute.AttributeRegistry;
@@ -35,8 +35,7 @@ import javafx.scene.control.TableColumn;
 import javafx.scene.layout.BorderPane;
 
 /**
- * ImportTableColumn extends a standard JavaFX table column to add field
- * validation.
+ * ImportTableColumn extends a standard JavaFX table column to add field validation.
  *
  * @author sirius
  */
@@ -78,13 +77,15 @@ public class ImportTableColumn extends TableColumn<TableRow, CellValue> {
                 attributeDescriptionClass = BooleanAttributeDescription.class;
             }
             try {
-                final AttributeDescription attributeDescription = attributeDescriptionClass.getDeclaredConstructor().newInstance();
+                final AttributeDescription attributeDescription = attributeDescriptionClass != null
+                        ? attributeDescriptionClass.getDeclaredConstructor().newInstance()
+                        : BooleanAttributeDescription.class.getDeclaredConstructor().newInstance();
+
                 for (final TableRow row : data) {
                     final CellValueProperty property = row.getProperty(columnIndex);
                     final String value = property.get().getOriginalText();
                     final String parsedValue = parser.translate(value, parserParameters);
                     final String errorMessage = attributeDescription.acceptsString(parsedValue);
-
                     columnFailed |= errorMessage != null;
                     if (parsedValue == null ? value == null : parsedValue.equals(value)) {
                         property.setText(value);
@@ -109,18 +110,15 @@ public class ImportTableColumn extends TableColumn<TableRow, CellValue> {
                 LOGGER.log(Level.SEVERE, ex.getLocalizedMessage(), ex);
             }
         } else {
-            for (final TableRow row : data) {
-                final CellValueProperty property = row.getProperty(columnIndex);
+            data.stream().map(row -> row.getProperty(columnIndex)).map(property -> {
                 property.setText(property.get().getOriginalText());
+                return property;
+            }).forEachOrdered(property -> {
                 property.setMessage(null, false);
-            }
+            });
         }
-
-        if (columnFailed) {
-            getGraphic().setStyle("-fx-background-color: rgba(255, 0, 0, 0.3);");
-        } else {
-            getGraphic().setStyle("-fx-background-color: transparent;");
-        }
+        getGraphic().setStyle(columnFailed ? "-fx-background-color: rgba(255, 0, 0, 0.3);"
+                : "-fx-background-color: transparent;");
     }
 
     public AttributeNode getAttributeNode() {
