+++ CoreImportExportPlugins/src/au/gov/asd/tac/constellation/plugins/importexport/ImportTopComponent.java
@@ -1,0 +1,70 @@
+/*
+ * Copyright 2010-2020 Australian Signals Directorate
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package au.gov.asd.tac.constellation.plugins.importexport;
+
+import au.gov.asd.tac.constellation.graph.Graph;
+import au.gov.asd.tac.constellation.views.JavaFxTopComponent;
+
+/**
+ * The Delimited Import Top Component. This class provides the Import Delimited window and handles all interactions with
+ * the graph.
+ *
+ * @author aldebaran30701
+ */
+public abstract class ImportTopComponent extends JavaFxTopComponent<ImportPane> {
+
+    public ImportTopComponent() {
+    }
+
+    @Override
+    protected void handleNewGraph(final Graph graph) {
+        preparePane();
+    }
+
+    @Override
+    protected void handleGraphOpened(final Graph graph) {
+        preparePane();
+    }
+
+    @Override
+    protected void handleGraphClosed(final Graph graph) {
+        preparePane();
+    }
+
+    protected abstract void preparePane();
+
+    /**
+     * This method is called from within the constructor to initialize the form. WARNING: Do NOT modify this code. The
+     * content of this method is always regenerated by the Form Editor.
+     */
+    // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
+    private void initComponents() {
+
+        javax.swing.GroupLayout layout = new javax.swing.GroupLayout(this);
+        this.setLayout(layout);
+        layout.setHorizontalGroup(
+            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGap(0, 400, Short.MAX_VALUE)
+        );
+        layout.setVerticalGroup(
+            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGap(0, 300, Short.MAX_VALUE)
+        );
+    }// </editor-fold>//GEN-END:initComponents
+
+    // Variables declaration - do not modify//GEN-BEGIN:variables
+    // End of variables declaration//GEN-END:variables
+}
