+++ CoreImportExportPlugins/src/au/gov/asd/tac/constellation/plugins/importexport/delimited/DelimitedImportTopComponent.java
@@ -15,8 +15,8 @@
  */
 package au.gov.asd.tac.constellation.plugins.importexport.delimited;
 
-import au.gov.asd.tac.constellation.graph.Graph;
-import au.gov.asd.tac.constellation.views.JavaFxTopComponent;
+import au.gov.asd.tac.constellation.plugins.importexport.ConfigurationPane;
+import au.gov.asd.tac.constellation.plugins.importexport.ImportTopComponent;
 import javafx.application.Platform;
 import org.openide.awt.ActionID;
 import org.openide.awt.ActionReference;
@@ -25,8 +25,8 @@ import org.openide.util.NbBundle.Messages;
 import org.openide.windows.TopComponent;
 
 /**
- * The Delimited Import Top Component. This class provides the Import Delimited
- * window and handles all interactions with the graph.
+ * The Delimited Import Top Component. This class provides the Import Delimited window and handles all interactions with
+ * the graph.
  *
  * @author aldebaran30701
  */
@@ -49,23 +49,34 @@ import org.openide.windows.TopComponent;
 @Messages({
     "CTL_ImportDelimitedFileAction=From Delimited File...",
     "HINT_ImportDelimitedFile=Import from Delimited File"})
-public final class DelimitedImportTopComponent extends JavaFxTopComponent<DelimitedImportPane> {
+public final class DelimitedImportTopComponent extends ImportTopComponent {
 
     private final DelimitedImportPane delimitedImportPane;
+    private final String helpText = "1. Click on the green plus icon to add files.\n"
+            + "2. Select your destination graph.\n"
+            + "3. Drag and drop attributes onto columns.\n"
+            + "4. Right click an attribute for more options.\n"
+            + "5. Click the 'Import' button to add data to your graph.\n"
+            + "6. Save your configuration using 'Options > Save'.\n\n"
+            + "HINTS:\n* See all supported attributes with 'Options > Show all schema attributes'.\n"
+            + "* Start typing to filter attributes (press delete to clear).";
+    protected final DelimitedImportController controller = new DelimitedImportController();
+    protected final ConfigurationPane configurationPane = new ConfigurationPane(controller, helpText);
+    protected final DelimitedSourcePane sourcePane = new DelimitedSourcePane(controller);
 
     public DelimitedImportTopComponent() {
+        super();
         setName(Bundle.CTL_ImportDelimitedFileAction());
         setToolTipText(Bundle.HINT_ImportDelimitedFile());
         initComponents();
-
-        delimitedImportPane = new DelimitedImportPane(this);
+        delimitedImportPane = new DelimitedImportPane(this, controller, configurationPane, sourcePane);
+        controller.setImportPane(delimitedImportPane);
         initContent();
     }
 
     @Override
     protected String createStyle() {
         return "resources/delimited-import.css";
-//        return null;
     }
 
     @Override
@@ -74,21 +85,7 @@ public final class DelimitedImportTopComponent extends JavaFxTopComponent<Delimi
     }
 
     @Override
-    protected void handleNewGraph(final Graph graph) {
-        preparePane();
-    }
-
-    @Override
-    protected void handleGraphOpened(final Graph graph) {
-        preparePane();
-    }
-
-    @Override
-    protected void handleGraphClosed(final Graph graph) {
-        preparePane();
-    }
-
-    private void preparePane() {
+    protected void preparePane() {
         Platform.runLater(() -> {
             delimitedImportPane.getSourcePane().updateDestinationGraphCombo();
             delimitedImportPane.updateSourcePane();
@@ -96,9 +93,8 @@ public final class DelimitedImportTopComponent extends JavaFxTopComponent<Delimi
     }
 
     /**
-     * This method is called from within the constructor to initialize the form.
-     * WARNING: Do NOT modify this code. The content of this method is always
-     * regenerated by the Form Editor.
+     * This method is called from within the constructor to initialize the form. WARNING: Do NOT modify this code. The
+     * content of this method is always regenerated by the Form Editor.
      */
     // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
     private void initComponents() {
