+++ CoreImportExportPlugins/src/au/gov/asd/tac/constellation/plugins/importexport/jdbc/ImportJDBCPlugin.java
@@ -218,7 +218,7 @@ public class ImportJDBCPlugin extends SimpleEditPlugin {
 
         int directedIx = ImportConstants.ATTRIBUTE_NOT_ASSIGNED_TO_COLUMN;
         for (int i = 0; i < transactionDefinitions.size(); i++) {
-            if (transactionDefinitions.get(i).getAttribute().getName().equals(ImportController.DIRECTED)) {
+            if (transactionDefinitions.get(i).getAttribute().getName().equals(JDBCImportController.DIRECTED)) {
                 directedIx = transactionDefinitions.get(i).getColumnIndex();
                 break;
             }
